[#3] Init transfer of methods to separate files
